+++ packages/hierarchies/src/hierarchies/HierarchyFiltering.ts
@@ -7,16 +7,91 @@ import { HierarchyNode, NonGroupingHierarchyNode } from "./HierarchyNode.js";
 import { HierarchyNodeIdentifier, HierarchyNodeIdentifiersPath } from "./HierarchyNodeIdentifier.js";
 import { GenericNodeKey, GroupingNodeKey, HierarchyNodeKey, InstancesNodeKey } from "./HierarchyNodeKey.js";
 
-/** @public */
+/**
+ * @public
+ * @deprecated in 1.7. Use `FilteringPathAutoExpandDepthInHierarchy` instead.
+ */
 export interface FilterTargetGroupingNodeInfo {
-  /** Key of the grouping node. */
+  /**
+   * Key of the grouping node.
+   * @deprecated in 1.6. Use `FilteringPathAutoExpandDepthInHierarchy` instead.
+   */
   key: GroupingNodeKey;
 
   /**
-   * Depth of the grouping node in the hierarchy.
-   * Generally, it can be retrieved from `parentKeys.length`.
+   * Depth up to which nodes in the hierarchy should be expanded.
+   * @deprecated in 1.6. Use `FilteringPathAutoExpandDepthInHierarchy` instead.
    */
   depth: number;
+}
+
+/**
+ * @public
+ * @deprecated in 1.7. Use `FilteringPathAutoExpandDepthInPath` or `FilteringPathAutoExpandDepthInHierarchy` instead.
+ */
+export interface FilteringPathAutoExpandOption {
+  /**
+   * Depth up to which nodes in the hierarchy should be expanded.
+   *
+   * If `includeGroupingNodes` is set to true, then depth should take into account the number of grouping nodes in hierarchy.
+   * @deprecated in 1.7. Use `FilteringPathAutoExpandDepthInPath` or `FilteringPathAutoExpandDepthInHierarchy` instead.
+   */
+  depth: number;
+  /**
+   * Whether or not `depth` includes grouping nodes.
+   *
+   * Use when you want to autoExpand only some of the grouping nodes.
+   *
+   * **Use case example:**
+   *
+   * You want to `autoExpand` only `Node1` and `GroupingNode1` in the following hierarchy:
+   * - Node1
+   *   - GroupingNode1
+   *     - GroupingNode2
+   *       - Element1
+   *       - Element2
+   * Then you provide `autoExpand: { depth: 2, includeGroupingNodes: true }`
+   *
+   * To get the correct depth use `HierarchyNode.parentKeys.length`.
+   * @deprecated in 1.7. Use `FilteringPathAutoExpandDepthInHierarchy` instead.
+   */
+  includeGroupingNodes?: boolean;
+}
+
+/** @public */
+export interface FilteringPathAutoExpandDepthInPath {
+  /**
+   * Depth that tells which nodes in the filtering path should be expanded.
+   *
+   * Use when you want to expand up to specific instance node and don't care about grouping nodes.
+   *
+   * **NOTE**: All nodes that are up to `depthInPath` will be expanded. Node at the `depthInPath` position won't be expanded.
+   */
+  depthInPath: number;
+}
+
+/** @public */
+export interface FilteringPathAutoExpandDepthInHierarchy {
+  /**
+   * Depth that tells which nodes in the hierarchy should be expanded.
+   *
+   * This should take into account the number of grouping nodes in hierarchy.
+   *
+   * * **Use case example:**
+   *
+   * You want to `autoExpand` only `Node1` and `GroupingNode1` in the following hierarchy:
+   * - Node1
+   *   - GroupingNode1
+   *     - GroupingNode2
+   *       - Element1
+   *       - Element2
+   * Then you provide `autoExpand: { depthInHierarchy: 2 }`
+   *
+   * To get the correct depth use `HierarchyNode.parentKeys.length`.
+   *
+   * **NOTE**: All nodes that are up to and including `depthInHierarchy` will be expanded.
+   */
+  depthInHierarchy: number;
 }
 
 /** @public */
@@ -27,8 +102,51 @@ export interface HierarchyFilteringPathOptions {
    * - If it's `true`, then all nodes up to the filter target will have `autoExpand` flag.
    * - If it's an instance of `FilterTargetGroupingNodeInfo`, then all nodes up to the grouping node that matches this property,
    * will have `autoExpand` flag.
+   * - If it's an instance of `FilteringPathAutoExpandOption`, then all nodes up to and including `depth` will have `autoExpand` flag.
+   * - If it's an instance of `FilteringPathAutoExpandDepthInPath`, then all nodes up to `depthInPath` will have `autoExpand` flag.
+   * - If it's an instance of `FilteringPathAutoExpandDepthInHierarchy`, then all nodes up to and including `depthInHierarchy` will have `autoExpand` flag.
    */
-  autoExpand?: boolean | FilterTargetGroupingNodeInfo;
+  autoExpand?:
+    | boolean
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    | FilterTargetGroupingNodeInfo
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    | FilteringPathAutoExpandOption
+    | FilteringPathAutoExpandDepthInHierarchy
+    | FilteringPathAutoExpandDepthInPath;
+}
+
+namespace HierarchyFilteringPathOptions {
+  export function mergeAutoExpandOptions(
+    lhs: HierarchyFilteringPathOptions["autoExpand"],
+    rhs: HierarchyFilteringPathOptions["autoExpand"],
+  ): HierarchyFilteringPathOptions["autoExpand"] {
+    if (rhs === true || lhs === true) {
+      return true;
+    }
+    if (!rhs || !lhs) {
+      return !!rhs ? rhs : lhs;
+    }
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    const lhsDepth = "depth" in lhs ? lhs.depth : "depthInPath" in lhs ? lhs.depthInPath : lhs.depthInHierarchy;
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    const rhsDepth = "depth" in rhs ? rhs.depth : "depthInPath" in rhs ? rhs.depthInPath : rhs.depthInHierarchy;
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    const isLhsDepthBasedOnPath = "depthInPath" in lhs ? true : !("key" in lhs) && !("depthInHierarchy" in lhs) && !lhs.includeGroupingNodes;
+    // eslint-disable-next-line @typescript-eslint/no-deprecated
+    const isRhsDepthBasedOnPath = "depthInPath" in rhs ? true : !("key" in rhs) && !("depthInHierarchy" in rhs) && !rhs.includeGroupingNodes;
+
+    if (isLhsDepthBasedOnPath) {
+      if (isRhsDepthBasedOnPath) {
+        return lhsDepth > rhsDepth ? lhs : rhs;
+      }
+      return lhs;
+    }
+    if (isRhsDepthBasedOnPath) {
+      return rhs;
+    }
+    return lhsDepth > rhsDepth ? lhs : rhs;
+  }
 }
 
 /**
@@ -58,9 +176,11 @@ export namespace HierarchyFilteringPath {
    *
    * For the `autoExpand` attribute, the merge chooses to auto-expand as deep as the deepest input:
    * - if any one of the inputs is `true`, return `true`,
-   * - else if both inputs are objects, return the one with the greater `depth` attribute.
-   * - else if one input is an object, return it.
-   * - else, return `false` or `undefined`.
+   * - else if only one of the inputs is an object, return it,
+   * - else if both inputs are falsy, return `false` or `undefined`,
+   * - else:
+   *    - if only one of the inputs has `includeGroupingNodes` set to `true` or `key` defined or `depthInHierarchy` set, return the other one,
+   *    - else return the one with greater `depth`, `depthInPath` or `depthInHierarchy`.
    *
    * @public
    */
@@ -73,18 +193,7 @@ export namespace HierarchyFilteringPath {
     }
 
     return {
-      autoExpand: ((): HierarchyFilteringPathOptions["autoExpand"] => {
-        if (rhs.autoExpand === true || lhs.autoExpand === true) {
-          return true;
-        }
-        if (typeof lhs.autoExpand === "object") {
-          if (typeof rhs.autoExpand === "object" && rhs.autoExpand.depth > lhs.autoExpand.depth) {
-            return rhs.autoExpand;
-          }
-          return lhs.autoExpand;
-        }
-        return rhs.autoExpand;
-      })(),
+      autoExpand: HierarchyFilteringPathOptions.mergeAutoExpandOptions(lhs.autoExpand, rhs.autoExpand),
     };
   }
 }
@@ -137,7 +246,7 @@ function extractFilteringPropsInternal(
  */
 export function createHierarchyFilteringHelper(
   rootLevelFilteringProps: HierarchyFilteringPath[] | undefined,
-  parentNode: Pick<NonGroupingHierarchyNode, "filtering"> | undefined,
+  parentNode: Pick<NonGroupingHierarchyNode, "filtering" | "parentKeys"> | undefined,
 ) {
   const filteringProps = extractFilteringPropsInternal(rootLevelFilteringProps, parentNode);
   const hasFilter = !!filteringProps;
@@ -186,7 +295,7 @@ export function createHierarchyFilteringHelper(
         | {
             pathMatcher: (identifier: HierarchyNodeIdentifier) => boolean;
           },
-    ): Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined => {
+    ): Pick<HierarchyNode, "autoExpand" | "filtering"> | undefined => {
       if (!hasFilter) {
         return undefined;
       }
@@ -203,7 +312,7 @@ export function createHierarchyFilteringHelper(
           reducer.accept(normalizedPath);
         }
       });
-      return reducer.getNodeProps();
+      return reducer.getNodeProps(parentNode);
     },
 
     /**
@@ -211,7 +320,7 @@ export function createHierarchyFilteringHelper(
      */
     createChildNodePropsAsync: (props: {
       pathMatcher: (identifier: HierarchyNodeIdentifier) => boolean | Promise<boolean>;
-    }): Promise<Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined> | Pick<HierarchyNode, "filtering" | "autoExpand"> | undefined => {
+    }): Promise<Pick<HierarchyNode, "autoExpand" | "filtering"> | undefined> | Pick<HierarchyNode, "autoExpand" | "filtering"> | undefined => {
       if (!hasFilter) {
         return undefined;
       }
@@ -236,11 +345,11 @@ export function createHierarchyFilteringHelper(
         reducer.accept(normalizedPath);
       }
       if (matchedPathPromises.length === 0) {
-        return reducer.getNodeProps();
+        return reducer.getNodeProps(parentNode);
       }
       return Promise.all(matchedPathPromises)
         .then((matchedPath) => matchedPath.forEach((normalizedPath) => normalizedPath && reducer.accept(normalizedPath)))
-        .then(() => reducer.getNodeProps());
+        .then(() => reducer.getNodeProps(parentNode));
     },
   };
 }
@@ -251,22 +360,49 @@ class MatchingFilteringPathsReducer {
   private _filteredChildrenIdentifierPaths = new Array<NormalizedFilteringPath>();
   private _isFilterTarget = false;
   private _filterTargetOptions = undefined as HierarchyFilteringPathOptions | undefined;
-  private _needsAutoExpand = false;
+  private _autoExpandOption: HierarchyFilteringPathOptions["autoExpand"] = false;
 
   public constructor(private _hasFilterTargetAncestor: boolean) {}
 
-  public accept({ path, options }: NormalizedFilteringPath) {
+  public accept(normalizedPath: NormalizedFilteringPath): void {
+    const { path, options } = normalizedPath;
     if (path.length === 1) {
       this._isFilterTarget = true;
       this._filterTargetOptions = HierarchyFilteringPath.mergeOptions(this._filterTargetOptions, options);
     } else if (path.length > 1) {
       this._filteredChildrenIdentifierPaths.push({ path: path.slice(1), options });
-      if (options?.autoExpand) {
-        this._needsAutoExpand = true;
-      }
+      this._autoExpandOption = HierarchyFilteringPathOptions.mergeAutoExpandOptions(options?.autoExpand, this._autoExpandOption);
     }
   }
-  public getNodeProps(): Pick<HierarchyNode, "filtering" | "autoExpand"> {
+
+  private getNeedsAutoExpand(parentNode: Pick<NonGroupingHierarchyNode, "parentKeys"> | undefined): boolean {
+    if (this._autoExpandOption === true) {
+      return true;
+    }
+    if (typeof this._autoExpandOption === "object") {
+      const parentLength = !parentNode
+        ? 0
+        : "key" in this._autoExpandOption ||
+            "depthInHierarchy" in this._autoExpandOption ||
+            // eslint-disable-next-line @typescript-eslint/no-deprecated
+            ("includeGroupingNodes" in this._autoExpandOption && this._autoExpandOption.includeGroupingNodes)
+          ? 1 + parentNode.parentKeys.length
+          : 1 + parentNode.parentKeys.filter((key) => !HierarchyNodeKey.isGrouping(key)).length;
+      const depth =
+        "depthInHierarchy" in this._autoExpandOption
+          ? this._autoExpandOption.depthInHierarchy
+          : "depth" in this._autoExpandOption
+            ? // eslint-disable-next-line @typescript-eslint/no-deprecated
+              this._autoExpandOption.depth
+            : // With `depthInPath` option we don't want to expand node that is at the `depthInPath` position
+              this._autoExpandOption.depthInPath - 1;
+
+      return parentLength < depth;
+    }
+    return false;
+  }
+
+  public getNodeProps(parentNode: Pick<NonGroupingHierarchyNode, "parentKeys"> | undefined): Pick<HierarchyNode, "autoExpand" | "filtering"> {
     return {
       ...(this._hasFilterTargetAncestor || this._isFilterTarget || this._filteredChildrenIdentifierPaths.length > 0
         ? {
@@ -277,7 +413,7 @@ class MatchingFilteringPathsReducer {
             },
           }
         : undefined),
-      ...(this._needsAutoExpand ? { autoExpand: true } : undefined),
+      ...(this.getNeedsAutoExpand(parentNode) ? { autoExpand: true } : undefined),
     };
   }
 }