+++ packages/hierarchies-react/CHANGELOG.md
@@ -1,5 +1,85 @@
 # @itwin/presentation-hierarchies-react
 
+## 1.9.2
+
+### Patch Changes
+
+- [#1039](https://github.com/iTwin/presentation/pull/1039): Bump iTwin.js core dependencies to `^5.1.1`.
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.7.2
+  - @itwin/presentation-shared@1.2.3
+  - @itwin/unified-selection@1.5.1
+
+## 1.9.1
+
+### Patch Changes
+
+- [#1033](https://github.com/iTwin/presentation/pull/1033): Fix unified selection tree state hooks `useUnifiedSelectionTree` and `useIModelUnifiedSelectionTree` not handling the recently added `createSelectableForGenericNode` prop.
+
+## 1.9.0
+
+### Minor Changes
+
+- [#1027](https://github.com/iTwin/presentation/pull/1027): Added `getActions` property to `TreeRenderer` and `TreeNodeRenderer` to allow providing custom actions for tree nodes.
+
+## 1.8.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.7.1
+
+## 1.8.0
+
+### Minor Changes
+
+- [#1019](https://github.com/iTwin/presentation/pull/1019): Add ability to customize `Selectable` objects that are added to unified selection, when a generic node in a unified selection-driven tree component is selected.
+
+  By default, when a generic node is selected in a unified selection-driven tree component, the `Selectable` that is created for it gets a per-tree unique identifier. This means that a similar generic node somewhere else in the tree or a different tree will not be considered the same. However, in some cases we do want generic nodes to be treated as the same, and for that matter unified selection-enabling tree state hooks now have a new `createSelectableForGenericNode` option:
+
+  ```tsx
+  import { useUnifiedSelectionTree } from "@itwin/presentation-hierarchies-react";
+
+  const treeState = useUnifiedSelectionTree({
+    selectionStorage,
+    createSelectableForGenericNode: useCallback<NonNullable<Props<typeof useUnifiedSelectionTree>["createSelectableForGenericNode"]>>(
+      (node, uniqueId) => ({ identifier: node.key.id, data: node, async *loadInstanceKeys() {} }),
+      [],
+    ),
+    // ...other options
+  });
+  ```
+
+### Patch Changes
+
+- Updated dependencies:
+  - @itwin/unified-selection@1.5.0
+
+## 1.7.2
+
+### Patch Changes
+
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.7.0
+
+## 1.7.1
+
+### Patch Changes
+
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.6.1
+
+## 1.7.0
+
+### Minor Changes
+
+- 64e80beaeb87e5797d8238c14fa9cd8dbfcee7c3: Added `abortSignal` to `getFilteredPaths` callback used by `UseTree` hook. This allows to cancel ongoing requests when component is unmounted or new request is started.
+
+### Patch Changes
+
+- Updated dependencies:
+  - @itwin/presentation-hierarchies@1.6.0
+
 ## 1.6.9
 
 ### Patch Changes