+++ packages/hierarchies/CHANGELOG.md
@@ -1,5 +1,70 @@
 # @itwin/presentation-hierarchies
 
+## 1.7.2
+
+### Patch Changes
+
+- [#1039](https://github.com/iTwin/presentation/pull/1039): Bump iTwin.js core dependencies to `^5.1.1`.
+- Updated dependencies:
+  - @itwin/presentation-shared@1.2.3
+
+## 1.7.1
+
+### Patch Changes
+
+- [#1024](https://github.com/iTwin/presentation/pull/1024): Update `depthInPath` documentation. Clarified that it does not set auto-expand for node that is at `depthInPath` position.
+
+## 1.7.0
+
+### Minor Changes
+
+- 8f7d200926b93e862276e1f978f6c891691e0dae: Refactored specifying the depth / level to which filtering path should be auto-expanded.
+
+  - Deprecated `FilteringPathAutoExpandOption` and `FilterTargetGroupingNodeInfo`.
+  - Added `FilteringPathAutoExpandDepthInPath` and `FilteringPathAutoExpandDepthInHierarchy` which should be used instead.
+
+  Migration:
+
+  ```ts
+  const pathExpandedToInstanceInPath: HierarchyFilteringPath = {
+    path: [instanceKey1, instanceKey2, instanceKey3],
+    options: {
+      // deprecated:
+      autoExpand: { depth: 1 },
+      // do this instead:
+      autoExpand: { depthInPath: 1 },
+    },
+  };
+
+  const pathExpandedToGroupingNode: HierarchyFilteringPath = {
+    path: [instanceKey1, instanceKey2, instanceKey3],
+    options: {
+      // deprecated:
+      autoExpand: { depth: 1, key: groupingNodeKey },
+      // also deprecated:
+      autoExpand: { depth: 1, includeGroupingNodes: true },
+      // do this instead:
+      autoExpand: { depthInHierarchy: 1 },
+    },
+  };
+  ```
+
+## 1.6.1
+
+### Patch Changes
+
+- [#1010](https://github.com/iTwin/presentation/pull/1010): Fixed `HierarchyFilteringPathOptions.autoExpand` depth option not working properly with grouping nodes. Added `includeGroupingNodes` attribute to `FilteringPathAutoExpandOption`, which allows auto-expanding only desired gouping nodes, when set together with `depth` value.
+
+## 1.6.0
+
+### Minor Changes
+
+- 69a2db67917f84e99d532e8e9aabbc02ec262d91: `HierarchyFilteringPathOptions.autoExpand` is now of type:
+  ```ts
+  autoExpand?: { depth: number } | boolean;
+  ```
+  When `depth` is set, only nodes up to specified `depth` in `HierarchyFilteringPath` will have `autoExpand` option set.
+
 ## 1.5.1
 
 ### Patch Changes